Add user save.

+++ app/src/main/java/io/github/gaeqs/quiz/data/User.java
@@ -29,13 +29,13 @@ public class User {
     private int maximumScore;
 
     @ColumnInfo(name = "maximum_score_time")
-    private int maximumScoreTime;
+    private long maximumScoreTime;
 
     @ColumnInfo(name = "last_played")
     private long lastPlayed;
 
 
-    public User(@NonNull String name, String imagePath, int matches, int maximumScore, int maximumScoreTime, long lastPlayed) {
+    public User(@NonNull String name, String imagePath, int matches, int maximumScore, long maximumScoreTime, long lastPlayed) {
         this.name = name;
         this.imagePath = imagePath;
         this.matches = matches;
@@ -80,11 +80,11 @@ public class User {
         this.maximumScore = maximumScore;
     }
 
-    public int getMaximumScoreTime() {
+    public long getMaximumScoreTime() {
         return maximumScoreTime;
     }
 
-    public void setMaximumScoreTime(int maximumScoreTime) {
+    public void setMaximumScoreTime(long maximumScoreTime) {
         this.maximumScoreTime = maximumScoreTime;
     }
